refactor to auth contoller
