+++ backend/server.js
@@ -3,11 +3,9 @@ const cors = require('cors');
 const bodyParser = require('body-parser');
 const mongoose = require('mongoose');
 const path = require('path');
-const bcrypt = require('bcrypt');
-const jwt = require('jsonwebtoken');
 const PDFDocument = require('pdfkit');
 const SVGtoPDF = require('svg-to-pdfkit');
-const User = require('./models/user');
+const auth = require('./controllers/Auth');
 
 PDFDocument.prototype.addSVG = function addSVG(svg, x, y, options) {
   return SVGtoPDF(this, svg, x, y, options);
@@ -65,7 +63,6 @@ function generateCard(w, h) {
   return `<svg id="preview" viewBox="0 0 1400 1400">${cells.join('')}</svg>`;
 }
 
-const SECRET = 'thisNeedsToChange';
 const DB_URL = process.env.MONGODB_URI || 'mongodb://localhost:27017/bingo';
 
 const server = express();
@@ -83,77 +80,11 @@ const db = mongoose.connection;
 db.on('error', console.error.bind(console, 'MongoDB connection error:'));
 
 server.use(express.static(path.join(__dirname, '../client/build')));
-server.post('/auth/register', async (req, res) => {
-  try {
-    const { username } = req.body;
-    let { password } = req.body;
 
-    if (!username || !password) {
-      res.json({
-        isValid: false,
-        errors: [
-          { message: 'Both username and password required' },
-        ],
-      });
-      return;
-    }
+server.post('/auth/register', auth.register);
+server.post('/auth/login', auth.login);
+server.post('/auth/reset', auth.reset);
 
-    if (await User.exists(username)) {
-      res.status(422).json({
-        isValid: false,
-        error: {
-          message: 'Username already exists.',
-        },
-      });
-      return;
-    }
-    password = await bcrypt.hash(password, 12);
-    const user = await User.create({ username, password });
-    const payload = { id: user._id, username: user.username };
-    const token = jwt.sign(payload, SECRET, { expiresInMinutes: '1h' });
-    res.json({ user, token });
-  } catch (e) {
-    res.json({ error: 'Failed to register' });
-  }
-});
-
-server.post('/auth/login', async (req, res) => {
-  const { username, password } = req.body;
-  if (!username || !password) return res.status(422).json({ error: 'Request must have both username and password.' });
-  const user = await User.authenticate(username, password);
-  const payload = { id: user._id, username: user.username };
-  const token = jwt.sign(payload, SECRET, { expiresIn: '1h' });
-  return user ? res.json({ user, token }) : res.status(403).json({ error: 'failed to authenticate' });
-});
-
-server.post('/auth/reset', async (req, res) => {
-  const {
-    username,
-    password,
-    confirmPassword,
-    confirmNewPassword,
-  } = req.body;
-
-  let { newPassword } = req.body;
-
-  if (password !== confirmPassword) {
-    res.status(422).send({ error: 'Current Passwords do not match.' });
-    return;
-  }
-  if (newPassword !== confirmNewPassword) {
-    res.status(422).send({ error: 'New Passwords do not match.' });
-    return;
-  }
-  const user = await User.authenticate(username, password);
-  if (!user) {
-    res.status(403).send({ error: 'Could not authenticate' });
-    return;
-  }
-  newPassword = await bcrypt.hash(newPassword, 12);
-  user.set({ password: newPassword });
-  const updatedUser = await user.save();
-  res.send({ user: updatedUser });
-});
 server.get('/cards', (req, res) => {
   res.json([]);
 });
